feat&fix: changes  - separate graphics.py on functions to plot the graphics and functions to get data for graphics  - fix functions to get property graphs
